Split award flow controller
At the moment we've got one controller for two separate award flows (the
award after initial assessment, and the award after verification). I've
renamed the first flow to verification to make it clear that it doesn't
guarantee an award after verification has been complete, so now we can
split the two flows in to separate controllers as their behaviours are
very different.
