+++ app/controllers/assessor_interface/assessments_controller.rb
@@ -45,8 +45,6 @@ module AssessorInterface
     delegate :application_form, to: :assessment
 
     def update_redirect_path(recommendation)
-      recommendation = "award" if recommendation == "verify"
-
       if recommendation == "request_further_information"
         [
           :preview,
