+++ app/views/assessor_interface/assessment_recommendation_verify/edit_reference_requests.html.erb
@@ -1,7 +1,7 @@
 <% content_for :page_title, "#{"Error: " if @form.errors.any?}#{t(".heading")}" %>
 <% content_for :back_link_url, assessor_interface_application_form_path(@application_form) %>
 
-<%= form_with model: @form, url: [:reference_requests, :assessor_interface, @application_form, @assessment, :assessment_recommendation_award] do |f| %>
+<%= form_with model: @form, url: [:reference_requests, :assessor_interface, @application_form, @assessment, :assessment_recommendation_verify] do |f| %>
   <%= f.govuk_error_summary %>
 
   <h1 class="govuk-heading-xl"><%= t(".heading") %></h1>
