+++ spec/system/assessor_interface/completing_assessment_spec.rb
@@ -71,14 +71,6 @@ RSpec.describe "Assessor completing assessment", type: :system do
     when_i_select_award_qts
     and_i_click_continue
     then_i_see_the(
-      :declare_assessment_recommendation_page,
-      application_id:,
-      assessment_id:,
-      recommendation: "award",
-    )
-
-    when_i_check_declaration
-    then_i_see_the(
       :reference_requests_assessment_recommendation_award_page,
       application_id:,
       assessment_id:,
@@ -99,14 +91,6 @@ RSpec.describe "Assessor completing assessment", type: :system do
     )
 
     when_i_send_the_teacher_email
-    then_i_see_the(
-      :confirm_assessment_recommendation_page,
-      application_id:,
-      assessment_id:,
-      recommendation: "award",
-    )
-
-    when_i_check_confirmation
     then_i_see_the(:assessor_application_status_page, application_id:)
 
     when_i_click_on_overview_button
